Delete redundant webhooks unit tests
- Delete webhook unit tests under controllers/api
- Improve webhook integration tests
  - Remove unnecessary cleanup
  - Remove unnecessary eventually statements
- Move webhook integration tests next to the code
