--- controllers/api/v1alpha1/integration/cfpackage_webhook_integration_test.go
@@ -1,104 +1,0 @@
-package integration_test
-
-import (
-	"context"
-
-	korifiv1alpha1 "code.cloudfoundry.org/korifi/controllers/api/v1alpha1"
-	. "code.cloudfoundry.org/korifi/controllers/controllers/workloads/testutils"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-)
-
-var _ = Describe("CFPackageMutatingWebhook Integration Tests", func() {
-	When("a CFApp record exists", func() {
-		const (
-			cfAppGUIDLabelKey = "korifi.cloudfoundry.org/app-guid"
-			cfPackageType     = "bits"
-			namespace         = "default"
-		)
-
-		var (
-			cfApp         *korifiv1alpha1.CFApp
-			cfAppGUID     string
-			cfPackageGUID string
-		)
-
-		BeforeEach(func() {
-			cfAppGUID = GenerateGUID()
-			cfPackageGUID = GenerateGUID()
-			cfApp = &korifiv1alpha1.CFApp{
-				TypeMeta: metav1.TypeMeta{
-					Kind:       "CFApp",
-					APIVersion: korifiv1alpha1.GroupVersion.Identifier(),
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      cfAppGUID,
-					Namespace: namespace,
-				},
-				Spec: korifiv1alpha1.CFAppSpec{
-					DisplayName:  "test-app",
-					DesiredState: "STOPPED",
-					Lifecycle: korifiv1alpha1.Lifecycle{
-						Type: "buildpack",
-					},
-				},
-			}
-			Expect(k8sClient.Create(context.Background(), cfApp)).To(Succeed())
-		})
-
-		AfterEach(func() {
-			Expect(k8sClient.Delete(context.Background(), cfApp)).To(Succeed())
-		})
-
-		When("a CFPackage record referencing the CFAPP is created", func() {
-			BeforeEach(func() {
-				cfPackage := &korifiv1alpha1.CFPackage{
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "CFPackage",
-						APIVersion: korifiv1alpha1.GroupVersion.Identifier(),
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      cfPackageGUID,
-						Namespace: namespace,
-					},
-					Spec: korifiv1alpha1.CFPackageSpec{
-						Type: cfPackageType,
-						AppRef: v1.LocalObjectReference{
-							Name: cfAppGUID,
-						},
-					},
-				}
-				Expect(k8sClient.Create(context.Background(), cfPackage)).To(Succeed())
-			})
-
-			AfterEach(func() {
-				cfPackage := &korifiv1alpha1.CFPackage{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      cfPackageGUID,
-						Namespace: namespace,
-					},
-				}
-				Expect(k8sClient.Delete(context.Background(), cfPackage)).To(Succeed())
-			})
-
-			It("should have CFAppGUID metadata label on it and its value should matches spec.appRef", func() {
-				cfPackageLookupKey := types.NamespacedName{Name: cfPackageGUID, Namespace: namespace}
-				createdCFPackage := new(korifiv1alpha1.CFPackage)
-
-				Eventually(func() map[string]string {
-					err := k8sClient.Get(context.Background(), cfPackageLookupKey, createdCFPackage)
-					if err != nil {
-						return nil
-					}
-					return createdCFPackage.ObjectMeta.Labels
-				}).ShouldNot(BeEmpty(), "CFPackage resource does not have any metadata.labels")
-
-				Expect(createdCFPackage.ObjectMeta.Labels).To(HaveKeyWithValue(cfAppGUIDLabelKey, cfAppGUID))
-			})
-		})
-	})
-})
